+++ app/test/e2e/queue-execution-concat.spec.js
@@ -174,6 +174,8 @@ describe('EXECUTION_CONCAT handling process', () => {
                             value.should.have.property('type').and.be.a('string').and.equal('dataset');
                         }
                     });
+                    content.should.have.property('file');
+                    message.fileUrl.should.include(content.file);
                 } else {
                     throw new Error(`Unexpected message type: ${content.type}`);
                 }
@@ -207,10 +209,15 @@ describe('EXECUTION_CONCAT handling process', () => {
                     case docImporterMessages.status.MESSAGE_TYPES.STATUS_READ_DATA:
                         content.should.have.property('id');
                         content.should.have.property('taskId').and.equal(message.taskId);
+                        content.should.have.property('hash').and.be.a('string');
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     case docImporterMessages.status.MESSAGE_TYPES.STATUS_READ_FILE:
                         content.should.have.property('id');
                         content.should.have.property('taskId').and.equal(message.taskId);
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     default:
                         throw new Error(`Unexpected message type: ${content.type}`);
@@ -361,6 +368,8 @@ describe('EXECUTION_CONCAT handling process', () => {
                                 value.should.have.property('type').and.be.a('string').and.equal('dataset');
                             }
                         });
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     default:
                         throw new Error(`Unexpected message type: ${content.type}`);
@@ -396,10 +405,15 @@ describe('EXECUTION_CONCAT handling process', () => {
                     case docImporterMessages.status.MESSAGE_TYPES.STATUS_READ_DATA:
                         content.should.have.property('id');
                         content.should.have.property('taskId').and.equal(message.taskId);
+                        content.should.have.property('hash').and.be.a('string');
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     case docImporterMessages.status.MESSAGE_TYPES.STATUS_READ_FILE:
                         content.should.have.property('id');
                         content.should.have.property('taskId').and.equal(message.taskId);
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     default:
                         throw new Error(`Unexpected message type: ${content.type}`);
@@ -576,6 +590,8 @@ describe('EXECUTION_CONCAT handling process', () => {
                                 value.should.have.property('type').and.be.a('string').and.equal('dataset');
                             }
                         });
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     default:
                         throw new Error(`Unexpected message type: ${content.type}`);
@@ -611,10 +627,15 @@ describe('EXECUTION_CONCAT handling process', () => {
                     case docImporterMessages.status.MESSAGE_TYPES.STATUS_READ_DATA:
                         content.should.have.property('id');
                         content.should.have.property('taskId').and.equal(message.taskId);
+                        content.should.have.property('hash').and.be.a('string');
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     case docImporterMessages.status.MESSAGE_TYPES.STATUS_READ_FILE:
                         content.should.have.property('id');
                         content.should.have.property('taskId').and.equal(message.taskId);
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     default:
                         throw new Error(`Unexpected message type: ${content.type}`);