+++ app/test/e2e/queue-execution-create.spec.js
@@ -179,6 +179,8 @@ describe('EXECUTION_CREATE handling process', () => {
                                 value.should.have.property('type').and.be.a('string').and.equal('dataset');
                             }
                         });
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     default:
                         throw new Error('Unexpected message type');
@@ -214,10 +216,15 @@ describe('EXECUTION_CREATE handling process', () => {
                     case docImporterMessages.status.MESSAGE_TYPES.STATUS_READ_DATA:
                         content.should.have.property('id');
                         content.should.have.property('taskId').and.equal(message.taskId);
+                        content.should.have.property('hash').and.be.a('string');
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     case docImporterMessages.status.MESSAGE_TYPES.STATUS_READ_FILE:
                         content.should.have.property('id');
                         content.should.have.property('taskId').and.equal(message.taskId);
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     default:
                         throw new Error('Unexpected message type');
@@ -367,6 +374,8 @@ describe('EXECUTION_CREATE handling process', () => {
                                 value.should.have.property('type').and.be.a('string').and.equal('dataset');
                             }
                         });
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     default:
                         throw new Error('Unexpected message type');
@@ -402,10 +411,15 @@ describe('EXECUTION_CREATE handling process', () => {
                     case docImporterMessages.status.MESSAGE_TYPES.STATUS_READ_DATA:
                         content.should.have.property('id');
                         content.should.have.property('taskId').and.equal(message.taskId);
+                        content.should.have.property('hash').and.be.a('string');
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     case docImporterMessages.status.MESSAGE_TYPES.STATUS_READ_FILE:
                         content.should.have.property('id');
                         content.should.have.property('taskId').and.equal(message.taskId);
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     default:
                         throw new Error('Unexpected message type');
@@ -709,6 +723,8 @@ describe('EXECUTION_CREATE handling process', () => {
                             value.should.have.property('type').and.be.a('string').and.equal('dataset');
                         }
                     });
+                    content.should.have.property('file');
+                    message.fileUrl.should.include(content.file);
                 } else {
                     throw new Error('Unexpected message type');
                 }
@@ -742,10 +758,15 @@ describe('EXECUTION_CREATE handling process', () => {
                     case docImporterMessages.status.MESSAGE_TYPES.STATUS_READ_DATA:
                         content.should.have.property('id');
                         content.should.have.property('taskId').and.equal(message.taskId);
+                        content.should.have.property('hash').and.be.a('string');
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     case docImporterMessages.status.MESSAGE_TYPES.STATUS_READ_FILE:
                         content.should.have.property('id');
                         content.should.have.property('taskId').and.equal(message.taskId);
+                        content.should.have.property('file');
+                        message.fileUrl.should.include(content.file);
                         break;
                     default:
                         throw new Error('Unexpected message type');